+++ packages/tsurlfilter/CHANGELOG.md
@@ -8,6 +8,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version changes -->
 <!-- e.g. [1.0.77]: https://github.com/AdguardTeam/tsurlfilter/compare/tsurlfilter-v1.0.76...tsurlfilter-v1.0.77 -->
 
+## Unreleased
+
+### Added
+
+- Optional property `maxNumberOfUnsafeRules` to `DeclarativeConverterOptions`
+  for unsafe rules (only for dynamic rulesets).
+
+### Changed
+
+- Updated [@adguard/agtree] to `v2.1.3`.
+- We throw error for empty modifier list in network rules.
+
+### Fixed
+
+- Handling missing children data in the deserializer for certain nodes.
+- URI encoded `$removeparam` value is not removed in MV3 [AdguardBrowserExtension#3014].
+- `$removeparam` fails to match encoded URL params in MV2 [AdguardBrowserExtension#3015].
+
+[AdguardBrowserExtension#3014]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3014
+
+## [3.0.7] - 2024-11-19
+
+### Fixed
+
+- `$removeparam` fails to match encoded URL params in MV2 [AdguardBrowserExtension#3015].
+
+[AdguardBrowserExtension#3015]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3015
+[3.0.7]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.0.7
+
+## [3.0.6] - 2024-11-02
+
+### Changed
+
+- Updated `@adguard/agtree` to `v2.1.3`.
+
+[3.0.6]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.0.6
+
 ## [3.1.0-alpha.7] - 2024-10-16
 
 ### Fixed